Add default value to app.name.
This was a strange bug I encountered only in acceptance tests.

Because `name` was undefined, when I tried to #deleteRecord, the
`{{input}}` helper in `new-app-card` component was trying to
`set` the name prop to "" (empty string) and it was undefined. I
got the following error:

    Uncaught Error: Attempted to handle event `didSetProperty` on <admin@model:user::ember578:null> while in state root.deleted.saved. Called with {name: name, oldValue: undefined, originalValue: undefined, value: }.

By giving it a default of empty string, it doesn't try to set on
the model, and no error is thrown.

This shouldn't happen, and it doesn't happen in the app when it's
running. Only the test.