[#29] Add app's location
Add's the apps location to the API response.

Basically we don't know what apps are in the database vs router. So when
the router boots we are going to match names to Rail's generated urls in
a map. An app instance will be able to lookup its url by it's name in
that map.

Fixes #29.