Use middleware to serve assets
Switch away from using ActionDispatch::Static with our own middleware
that just wraps Static.

Reason being that ActionDispatch::Static will check to see if there is a
file on disk for each request. If there is a file, it will serve it,
otherwise it will continue down the middleware stack. This is great for
development mode, but is bad for a high req/s production environment if
every http request needs to check disk.

This is a simple fix to that problem. Since we want to use the
middleware to serve our assets, but don't want to do a ``File.exists?``
call every request we will first verify that the request is for a
front end build admin area asset. If it is, we handoff to
ActionDispatch::Static, otherwise continue down the middleware stack.

This should give us the best of both worlds.