Remove all mentions of DSA keys as they don't work.
So here's the thing...

This commits removes DSA bits from FrontEndBuilds::Utils::SSHPubKeyConvert as this app doesn't support DSA keys.

If we wanted to in the future we would have to modify the PubKey#verify
method.

I looked all around the internet for a gem or ruby code to convert an
SSH RSA key to a x509 and couldn't find any code but the snippet already
in this codebase.

I've removed the extra DSA code and modfied the README and specs to
match